feat: puts components into their own files and folders, gives them scss, and routes the pages properly. Lesson_25 -JosephCaballero
