+++ lesson_25/Josephs_template/src/components/programs/Programs.scss
@@ -1,0 +1,7 @@
+
+  .program{
+    display: flex;
+    justify-content: center;
+    width: 50%;
+    flex-direction: column;
+  }
