+++ lesson_25/Josephs_template/src/components/programs/Programs.tsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import './Programs.scss';
+
+interface ProgramProps {
+  title: string;
+  description: string;
+}
+
+export const Programs: React.FC<ProgramProps> = ({title, description}) => {
+  return (
+    <>
+      <li className="program">
+        <h3>{title}</h3>
+        <p>{description}</p>
+      </li>
+    </>
+  );
+};
