+++ lesson_25/Josephs_template/src/components/programsList/ProgramsList.scss
@@ -1,0 +1,8 @@
+
+.programs {
+    display: flex;
+    flex-flow: row wrap;
+    list-style: none;
+    padding: 0;
+    justify-content: space-between;
+  }
