+++ lesson_25/Josephs_template/src/components/programsList/ProgramsList.tsx
@@ -1,0 +1,23 @@
+import './ProgramsList.scss';
+import React from 'react';
+
+import {useProgramContext} from '../../contexts/ProgramContext';
+import {Programs} from '../programs/Programs';
+
+const ProgramsList: React.FC = () => {
+  const {programs} = useProgramContext();
+  return (
+    <>
+      <ul className="programs">
+        {programs.map((each, index) => (
+          <Programs
+            key={index}
+            title={each.title}
+            description={each.description}
+          />
+        ))}
+      </ul>
+    </>
+  );
+};
+export default ProgramsList;
