+++ lesson_25/Josephs_template/src/main.tsx
@@ -1,16 +1,14 @@
 import App from './App.tsx';
-import { Home } from './pages/Home/Home.tsx';
-import { OtherPage, ProgramProvider } from './pages/Home/Home.tsx';
-import ErrorPage from './routes/errorPage.tsx';
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import {Home} from './pages/Home/Home.tsx';
+import {OtherPage} from './pages/OtherPage/OtherPage.tsx';
+import ErrorPage from './routes/ErrorPage.tsx';
+import {QueryClient, QueryClientProvider} from '@tanstack/react-query';
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import { RouterProvider, createBrowserRouter } from 'react-router-dom';
-
-
+import {RouterProvider, createBrowserRouter} from 'react-router-dom';
 
 import './index.scss';
-
+import { ProgramProvider } from './contexts/ProgramContext.tsx';
 
 const queryClient = new QueryClient();
 
@@ -25,14 +23,14 @@ const router = createBrowserRouter([
         element: <Home />,
         errorElement: <ErrorPage />,
       },
+      {
+        path: '/otherPage',
+        element: <OtherPage />,
+        errorElement: <ErrorPage />,
+      },
     ],
   },
-  {
-    path: '/otherPage',
-    element: <OtherPage />,
-    errorElement: <ErrorPage />,
-
-  },
+  {},
 ]);
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
