+++ lesson_25/Josephs_template/src/pages/Home/Home.scss
@@ -59,20 +59,6 @@ article {
   flex-direction: column;
   max-width: 1280px;
   margin: auto;
-
-  .programs {
-    display: flex;
-    flex-flow: row wrap;
-    list-style: none;
-    padding: 0;
-    justify-content: space-between;
-  }
-}
-
-.program {
-  display: flex;
-  flex-direction: column;
-  max-width: 48%;
 }
 
 @media screen and (max-width: 1080px) {
